chore(PPDSC-2329): chore-ppdse-2329-refactored
